% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_base.R
\name{exceedance_profile}
\alias{exceedance_profile}
\title{Profile the data}
\usage{
exceedance_profile(x, params, ...)
}
\arguments{
\item{x}{numeric, p-values from some tests
(E.g T-test for high-thoughput gene data.), with each element representing a
hypothesis.}

\item{params}{exceedance_parameters object}

\item{...}{not used}
}
\value{
an exceedance_profile object
}
\description{
Preprocessing the data and collecting information before performing
the exceedance control. This step is developed for reducing the
computational burder only. It is not related to statistics.
}
\examples{
## The 3rd pvalue statistic
params <- param_fast_GW(statistic = "kth_p", param1 = 3)

## generate p-values
x <- rbeta(10, 1, 10)

## profile the data
profile <- profile_pvalue(x,params)


## compute the 95\% confidence envolop
alpha <- 0.05

## reject the first three hypotheses
exceedance_bound(profile, alpha, ri = 3)

## reject the hypothese which pvalues are equal to
## the first three samples.
## In other word, this is equivalent to reject the first three hypotheses
exceedance_bound(profile, alpha, rx = x[1:3])

## reject the hypotheses which have the lowest 3 p-values
exceedance_bound(profile, alpha, sri = 3)


## Determine which hypotheses can be rejected while controlling the
## exceedance rate: P(FDP > bound) < alpha
alpha <- 0.05
bound <- 0.2
exceedance_inference(profile, alpha, bound)

}
